split facebook and telegram
